[20] Move functions to s/options.go
